AWS/aula02/pyproject.toml
[tool.poetry]
name = "aula02"
version = "0.1.0"
description = ""
authors = ["Thiago Ferraz <thiagovdferraz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.5"
boto3 = "^1.37.11"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
